INTERNAL: Extract typed value operations to typed_ops.h
